MATC: Avoid deleting variable in nested calls to `evalclause`.
`evalclause` can be called recursively for "for" statements. In this
case, variables that are defined in the outer call might be updated in
the inner call. Deleting the variable in the inner call invalidates the
variable in the outer call.

Avoid that by checking if a variable exists and only creating it if it
doesn't.

Also synchronize `matc/src/eval.c` to `post/matc/src/eval.c` (only
keeping different licenses as they are). Also synchronize the
`case forsym:` in `ElmerGUI/matc/src/eval.c`.